guestInfo, Booking 그리고 ResetPassword 확인
